Refactored the module name and structure
