+++ go.mod
@@ -1,9 +1,5 @@
-module src/lob
+module github.com/dacero/labyrinth-of-babel
 
 go 1.15
 
-require (
-	github.com/go-sql-driver/mysql v1.5.0
-	github.com/onsi/ginkgo v1.15.0
-	github.com/onsi/gomega v1.10.5
-)
+require github.com/go-sql-driver/mysql v1.5.0
